+++ src/app/service/poap/EndPOAP.ts
@@ -9,8 +9,9 @@ import POAPUtils, { POAPFileParticipant } from '../../utils/POAPUtils';
 import { CommandContext } from 'slash-create';
 import Log from '../../utils/Log';
 
-export default async (ctx: CommandContext, guildMember: GuildMember): Promise<any> => {
+export default async (guildMember: GuildMember, ctx?: CommandContext): Promise<any> => {
 	const db: Db = await dbInstance.dbConnect(constants.DB_NAME_DEGEN);
+	
 	await POAPUtils.validateUserAccess(guildMember, db);
 	
 	const poapSettingsDB: Collection = db.collection(constants.DB_COLLECTION_POAP_SETTINGS);
@@ -44,8 +45,10 @@ export default async (ctx: CommandContext, guildMember: GuildMember): Promise<an
 	const listOfParticipants = await POAPUtils.getListOfParticipants(guildMember, db, channel);
 	
 	if (listOfParticipants.length <= 0) {
-		await guildMember.send({ content: `Event ended. No participants found for ${channel.name} in ${channel.guild.name}.` });
-		await ctx.send(`Hey ${ctx.user.mention}, I just sent you a DM!`);
+		await guildMember.send({ content: `Event ended. No participants found for \`${channel.name}\` in \`${channel.guild.name}\`.` });
+		if (ctx) {
+			await ctx.send(`Hey ${ctx.user.mention}, I just sent you a DM!`);
+		}
 		return;
 	}
 	
@@ -53,13 +56,33 @@ export default async (ctx: CommandContext, guildMember: GuildMember): Promise<an
 	const currentDate = (new Date()).toISOString();
 	const fileName = `${channel.guild.name}_${channel.name}_${listOfParticipants.length}_participants.csv`;
 	await guildMember.send({
-		content: `Total Participants: \`${listOfParticipants.length} participants\`\n` +
-			`Guild: \`${channel.guild.name}\`\n` +
-			`Channel: \`${channel.name}\`\n` +
-			`Date: \`${currentDate} UTC\`.`,
+		embeds: [
+			{
+				title: 'POAP Distribution Results',
+				fields: [
+					{ name: 'Date', value: `${currentDate} UTC`, inline: true },
+					{ name: 'Event', value: `${poapSettingsDoc.event}`, inline: true },
+					{ name: 'Discord Server', value: channel.guild.name, inline: true },
+					{ name: 'Location', value: channel.name, inline: true },
+					{ name: 'Total Participants', value: `${listOfParticipants.length}`, inline: true },
+				],
+			},
+		],
 		files: [{ name: fileName, attachment: bufferFile }],
 	});
-	await ctx.send(`Hey ${ctx.user.mention}, I just sent you a DM!`);
+	Log.info('POAPs distributed', {
+		indexMeta: true,
+		meta: {
+			guildId: guildMember.guild.id,
+			guildName: guildMember.guild.name,
+			totalParticipants: listOfParticipants.length,
+			location: channel.name,
+		},
+	});
+
+	if (ctx) {
+		await ctx.send(`Hey ${ctx.user.mention}, I just sent you a DM!`);
+	}
 
 	if (guildMember.id !== guildMember.user.id) {
 		return guildMember.send({ content: `Previous event ended for <@${guildMember.id}>.` });